+++ zellij-server/src/screen.rs
@@ -751,6 +751,7 @@ pub enum ScreenInstruction {
         stacked_pane_list: bool,
         default_editor: Option<PathBuf>,
         advanced_mouse_actions: bool,
+        mouse_scroll_resize: bool,
         mouse_hover_effects: bool,
         visual_bell: bool,
         focus_follows_mouse: bool,
@@ -1458,6 +1459,7 @@ pub(crate) struct Screen {
     web_sharing: WebSharing,
     current_pane_group: Rc<RefCell<PaneGroups>>,
     advanced_mouse_actions: bool,
+    mouse_scroll_resize: bool,
     mouse_hover_effects: bool,
     visual_bell: bool,
     focus_follows_mouse: bool,
@@ -1560,6 +1562,7 @@ impl Screen {
         web_clients_allowed: bool,
         web_sharing: WebSharing,
         advanced_mouse_actions: bool,
+        mouse_scroll_resize: bool,
         mouse_hover_effects: bool,
         visual_bell: bool,
         focus_follows_mouse: bool,
@@ -1620,6 +1623,7 @@ impl Screen {
             current_pane_group: Rc::new(RefCell::new(current_pane_group)),
             currently_marking_pane_group: Rc::new(RefCell::new(HashMap::new())),
             advanced_mouse_actions,
+            mouse_scroll_resize,
             mouse_hover_effects,
             visual_bell,
             focus_follows_mouse,
@@ -3348,6 +3352,7 @@ impl Screen {
             self.current_pane_group.clone(),
             self.currently_marking_pane_group.clone(),
             self.advanced_mouse_actions,
+            self.mouse_scroll_resize,
             self.mouse_hover_effects,
             self.focus_follows_mouse,
             self.mouse_click_through,
@@ -5042,6 +5047,7 @@ impl Screen {
         stacked_pane_list: bool,
         default_editor: Option<PathBuf>,
         advanced_mouse_actions: bool,
+        mouse_scroll_resize: bool,
         mouse_hover_effects: bool,
         visual_bell: bool,
         focus_follows_mouse: bool,
@@ -5067,6 +5073,7 @@ impl Screen {
         self.copy_options.copy_on_select = copy_on_select;
         self.pane_frame_style = pane_frame_style;
         self.advanced_mouse_actions = advanced_mouse_actions;
+        self.mouse_scroll_resize = mouse_scroll_resize;
         self.mouse_hover_effects = mouse_hover_effects;
         self.visual_bell = visual_bell;
         self.focus_follows_mouse = focus_follows_mouse;
@@ -5094,6 +5101,7 @@ impl Screen {
             tab.set_pane_frames(pane_frame_style);
             tab.update_arrow_fonts(should_support_arrow_fonts);
             tab.update_advanced_mouse_actions(advanced_mouse_actions);
+            tab.update_mouse_scroll_resize(mouse_scroll_resize);
             tab.update_mouse_hover_effects(mouse_hover_effects);
             tab.update_focus_follows_mouse(focus_follows_mouse);
             tab.update_mouse_click_through(mouse_click_through);
@@ -6228,6 +6236,7 @@ pub(crate) fn screen_thread_main(
         .unwrap_or(false);
     let web_sharing = config_options.web_sharing.unwrap_or_else(Default::default);
     let advanced_mouse_actions = config_options.advanced_mouse_actions.unwrap_or(true);
+    let mouse_scroll_resize = config_options.mouse_scroll_resize.unwrap_or(true);
     let mouse_hover_effects = config_options.mouse_hover_effects.unwrap_or(true);
     let visual_bell = config_options.visual_bell.unwrap_or(true);
     let focus_follows_mouse = config_options.focus_follows_mouse.unwrap_or(false);
@@ -6270,6 +6279,7 @@ pub(crate) fn screen_thread_main(
         web_clients_allowed,
         web_sharing,
         advanced_mouse_actions,
+        mouse_scroll_resize,
         mouse_hover_effects,
         visual_bell,
         focus_follows_mouse,
@@ -9459,6 +9469,7 @@ pub(crate) fn screen_thread_main(
                 stacked_pane_list,
                 default_editor,
                 advanced_mouse_actions,
+                mouse_scroll_resize,
                 mouse_hover_effects,
                 visual_bell,
                 focus_follows_mouse,
@@ -9484,6 +9495,7 @@ pub(crate) fn screen_thread_main(
                         stacked_pane_list,
                         default_editor,
                         advanced_mouse_actions,
+                        mouse_scroll_resize,
                         mouse_hover_effects,
                         visual_bell,
                         focus_follows_mouse,